created business logic for create account

+++ accounts/src/main/java/com/banking/accounts/entity/Accounts.java
@@ -12,7 +12,7 @@ import lombok.*;
 @NoArgsConstructor
 public class Accounts extends BaseEntity {
 
-    private int customerId;
+    private Long customerId;
 
     @Id
     private long accountNumber;
